% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ERvar_split.R
\name{ERvar_split}
\alias{ERvar_split}
\title{Extremely randomized split}
\usage{
ERvar_split(X, Y, V, qui, type, ntry = 3, timeScale)
}
\arguments{
\item{ntry}{}
}
\description{
Extremely randomized split
}
\keyword{internal}
