% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OOB_rf.R
\name{OOB.rfshape}
\alias{OOB.rfshape}
\title{OOB predictions and error for Extra Frechet rf}
\usage{
OOB.rfshape(
  rf,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Shape = NULL,
  Image = NULL,
  Y,
  timeScale
)
}
\arguments{
\item{Y}{}
}
\description{
OOB predictions and error for Extra Frechet rf
}
\keyword{internal}
