% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_FT.R
\name{pred.FT}
\alias{pred.FT}
\title{Predict Frechet tree}
\usage{
pred.FT(
  tree,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Image = NULL,
  Curve.app = NULL,
  Factor.app = NULL,
  Image.app = NULL,
  timeScale
)
}
\arguments{
\item{tree}{: Frechet tree.}

\item{Curve}{[list]: A list that contains the input curves.}

\item{Scalar}{[list]: A list that contains the input scalars.}

\item{Factor}{[list]: A list that contains the input factors.}

\item{Image}{[list]: A list that contains the input images.}

\item{Curve.app}{[list]: A list that contains the learning input curves.}

\item{Factor.app}{[list]: A list that contains the learning input factors.}

\item{Image.app}{[list]: A list that contains the learning input images.}

\item{timeScale}{[numeric]: Time scale for the input and output curves (\code{timeScale=0.1} by default)}
}
\description{
Predict Frechet tree
}
