% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{predict.ExtraFrech}
\alias{predict.ExtraFrech}
\title{Predict with Extra Frechet random forests}
\usage{
\method{predict}{ExtraFrech}(
  object,
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Image = NULL,
  Curve.app = NULL,
  Factor.app = NULL,
  Image.app = NULL,
  ...
)
}
\arguments{
\item{object}{: Frechet random forest}

\item{Curve}{[list]: bla bla}

\item{Scalar}{[list]: bla bla}

\item{Factor}{[list]: bla bla bla}

\item{Image}{[list]: bla bla}

\item{Curve.app}{[list]: bla bla}

\item{Factor.app}{[list]: bla bla}

\item{Image.app}{[list]: bla bla}

\item{...}{: bla bla}
}
\description{
Predict with Extra Frechet random forests
}
