% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtraFrech.R
\name{rf_para}
\alias{rf_para}
\title{Parallelized Frechet random Forest}
\usage{
rf_para(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Image = NULL,
  Y,
  mtry,
  ntree,
  ncores,
  ntry = 3,
  timeScale
)
}
\arguments{
\item{ntry}{}
}
\description{
Parallelized Frechet random Forest
}
\keyword{internal}
