% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rtmax.R
\name{Rtmax}
\alias{Rtmax}
\title{Randomized Frechet tree}
\usage{
Rtmax(
  Curve = NULL,
  Scalar = NULL,
  Factor = NULL,
  Image = NULL,
  Y,
  mtry,
  ntry = 3,
  timeScale,
  ...
)
}
\arguments{
\item{Curve}{[list]:}

\item{Scalar}{[list]:}

\item{Factor}{[list]:}

\item{Image}{[list]:}

\item{Y}{[list]:}

\item{mtry}{[integer]:}

\item{ntry}{[numeric]:}

\item{timeScale}{[numeric]:}

\item{...}{: option}
}
\description{
Randomized Frechet tree
}
\keyword{internal}
